+++ lib/runners/processor/cppcheck.rb
@@ -1,7 +1,9 @@
 module Runners
   class Processor::Cppcheck < Processor
+    include CPlusPlus
+
     Schema = StrongJSON.new do
-      let :runner_config, Schema::BaseConfig.base.update_fields { |fields|
+      let :runner_config, Schema::BaseConfig.cplusplus.update_fields { |fields|
         fields.merge!(
           target: enum?(string, array(string)),
           ignore: enum?(string, array(string)),
@@ -102,6 +104,7 @@ module Runners
         *ignore,
         *project,
         *language,
+        *config_include_path,
         *args,
         *target
       )