+++ lib/runners/schema/config.rb
@@ -25,6 +25,13 @@ module Runners
         )
       })
 
+      let :cplusplus, (base.update_fields { |fields|
+        fields.merge!(
+          apt: enum?(string, array(string)),
+          'include-path': enum?(string, array(string))
+        )
+      })
+
       let :npm_install, enum(boolean, literal('development'), literal('production'))
 
       let :npm, (base.update_fields { |fields|