+++ test/smokes/cppcheck/config_string_values/myinclude/テスト.hpp
@@ -1,0 +1,5 @@
+#include <stdio.h>
+#include <libfastjson/json.h>
+#include <stdlib.h>
+
+#define N 31