+++ test/smokes/cppcheck/config_string_values/src/test.c
@@ -1,0 +1,10 @@
+#include "テスト.hpp"
+
+#define MY_N 31
+
+void foo()
+{
+  char buf[N];
+  buf[MY_N - 1] = 0;
+  buf[MY_N] = 0;
+}