+++ test/smokes/cppcheck/include_path/source/baz/myinclude/example1.h
@@ -1,0 +1,6 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <libfastjson/json.h>
+
+#define M 11