+++ test/smokes/cppcheck/include_path/source/baz/myinclude/example2.h
@@ -1,0 +1,5 @@
+#include <string.h>
+#include <gdbm.h>
+
+#define N2 (2*M)
+#define N3 (3*M)