+++ test/smokes/cppcheck/include_path/source/myinc/example.h
@@ -1,0 +1,6 @@
+#include <stdio.h>
+#include <readline/readline.h>
+#include <stdlib.h>
+#include <string.h>
+
+#define N 21