+++ test/smokes/cppcheck/include_path_default/source/foo/テスト.cxx
@@ -1,0 +1,11 @@
+#include <gdbm.h>
+#include "example.hpp"
+#include <stdio.h>
+
+int main()
+{
+    char buf[N];
+    buf[25] = 0;
+    buf[24] = 0;
+    return 0;
+}