+++ test/smokes/cppcheck/include_path_default/source/myinc/a/inc_hh.hh
@@ -1,0 +1,3 @@
+#include <stdio.h>
+#include "inc_hxx.HXX"
+#include <ctype.h>