+++ test/smokes/cppcheck/include_path_default/source/test.c
@@ -1,0 +1,12 @@
+#include "baz/myinclude/example1.H"
+#include <readline/readline.h>
+#include "example2.H++"
+#include "test.h"
+
+int foo()
+{
+  char buf[N];
+  buf[16] = 0;
+  buf[17] = 0;
+  return 0;
+}