+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - **stylelint** Add `stylelint-order` to pre-installed list [#1310](https://github.com/sider/runners/pull/1310)
 - Optimize git-blame [#1312](https://github.com/sider/runners/pull/1312)
 - Remove unused `Result::MissingFilesFailure` [#1315](https://github.com/sider/runners/pull/1315)
+- **Cppcheck** Support `-I` option to handle include paths [#1314](https://github.com/sider/runners/pull/1314)
 
 ## 0.30.0